+++ lib/pouch_io.dart
@@ -3,3 +3,4 @@ library pouch_io;
 
 export 'src/io.dart';
 export 'src/path_util.dart';
+export 'src/require_file.dart';