+++ lib/src/bool_extensions.dart
@@ -1,0 +1,9 @@
+/// Extensions to [bool]
+///
+/// `since 0.7.0`
+extension BoolExtenstions on bool {
+  /// Returns `true` when [this] is `false`
+  ///
+  /// `since 0.7.0`
+  bool get isFalse => this == false;
+}